requirements.txt
typing~=3.7.4.3
fedot==0.7.0
thegolem==0.3.1
scikit-learn~=1.2.2
opencv-python
numpy~=1.23.4
pandas~=1.4.3
matplotlib~=3.5.3
psutil~=5.9.3
setuptools~=67.8.0
tabulate~=0.9.0
datasets
# torchcnnbuilder~=0.1.0
torch # Should be installed before requirements.txt file application if you need CUDA support